Refactored lib.rs into separate files in the src directory
